[Notification] moves desktop tool in user account
