+++ docs/sections/plugins/img/notification_template.jpg
