+++ docs/sections/plugins/img/notifications_dropdown.jpg
