+++ docs/sections/plugins/img/notifications_mockup.png
