+++ docs/sections/plugins/img/notifications_system_uml.jpg
