+++ src/main/core/Resources/modules/account/routing.js
@@ -1,0 +1,8 @@
+
+function route(sectionName) {
+  return `/account/${sectionName}`
+}
+
+export {
+  route
+}
