+++ src/plugin/notification/Resources/config/services.yml
@@ -4,12 +4,13 @@ imports:
     - { resource: services/controller.yml }
     - { resource: services/listener.yml }
     - { resource: services/manager.yml }
+    - { resource: services/updater.yml }
 
 services:
     _defaults:
-        autowire: false #change to true
+        autowire: false
         autoconfigure: false
-        public: true #because controller
+        public: true
 
     Icap\NotificationBundle\Configuration\DefaultNotificationPlatformConfiguration:
         tags: ['claroline.configuration']
