+++ src/plugin/notification/Resources/modules/account/notifications/store/index.js
@@ -1,0 +1,9 @@
+import {actions} from '#/plugin/notification/account/notifications/store/actions'
+import {reducer} from '#/plugin/notification/account/notifications/store/reducer'
+import {selectors} from '#/plugin/notification/account/notifications/store/selectors'
+
+export {
+  actions,
+  reducer,
+  selectors
+}
