--- src/plugin/notification/Resources/modules/tools/notification/index.js
@@ -1,8 +1,0 @@
-import {reducer} from '#/plugin/notification/tools/notification/store'
-import {NotificationTool} from '#/plugin/notification/tools/notification/containers/tool'
-
-export default {
-  component: NotificationTool,
-  store: reducer,
-  styles: []
-}
