--- src/plugin/notification/Resources/modules/tools/notification/store/index.js
@@ -1,9 +1,0 @@
-import {actions} from '#/plugin/notification/tools/notification/store/actions'
-import {reducer} from '#/plugin/notification/tools/notification/store/reducer'
-import {selectors} from '#/plugin/notification/tools/notification/store/selectors'
-
-export {
-  actions,
-  reducer,
-  selectors
-}
