+++ src/plugin/notification/Resources/translations/notification.fr.json
@@ -5,8 +5,5 @@
     "more_unread": "+ %count% autres nouvelles notifications",
     "done_at": "le %date%",
     "disable_notifications": "Ne pas suivre cette ressource",
-    "disable_notifications_info": "Ne pas être notifié des nouveautés de cette ressource",
-    "enable_notifications": "Suivre cette ressource",
-    "enable_notifications_info": "Être notifié des nouveautés de cette ressource",
-    "mark_all_viewed": "Tout marquer comme lu"
+    "enable_notifications": "Suivre cette ressource"
 }
