--- src/plugin/notification/Resources/translations/tools.en.json
@@ -1,3 +1,0 @@
-{
-    "notification": "Notifications"
-}
