--- src/plugin/notification/Resources/translations/tools.fr.json
@@ -1,3 +1,0 @@
-{
-    "notification": "Notifications"
-}
